Fixed a few warnings.
Rev: src/global.h:1.62
Rev: src/interpret.c:1.204
Rev: src/interpret_functions.h:1.58

+++ src/global.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: global.h,v 1.61 2001/06/04 17:43:22 mast Exp $
+ * $Id: global.h,v 1.62 2001/06/10 00:22:47 grubba Exp $
  */
 #ifndef GLOBAL_H
 #define GLOBAL_H
@@ -203,7 +203,7 @@ void *alloca();
 #endif
 
 #define MAX_INT32 2147483647
-#define MIN_INT32 -2147483648
+#define MIN_INT32 -(INT32)2147483648
 
 #define INT16 short
 #define INT8 char
